DSS-971: Fix assertion.
git-svn-id: https://svn.cesecore.eu/svn/signserver-restricted/trunk@6835 91c8134f-2a44-0410-ba45-dce2f532dbc6

+++ signserver/modules/SignServer-Test-P11/src/test/java/org/signserver/server/cryptotokens/P11SignTest.java
@@ -952,7 +952,8 @@ public class P11SignTest extends ModulesTestCase {
 
             final RSAPublicKey pubKey = (RSAPublicKey) getPublicKeyFromRequest(req);
             
-            assertEquals("Returned public exponent", 5, pubKey.getPublicExponent());
+            assertEquals("Returned public exponent",
+                    BigInteger.valueOf(5), pubKey.getPublicExponent());
         } finally {
             try {
                 workerSession.removeKey(workerId, TEST_KEY_ALIAS);
